pyproject.toml
[tool.poetry]
name = "pylottie"
version = "2021"
description = "Convert .tgs and .lottie to .webp or .gif using pyppeteer."
authors = ["FredHappyface"]
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"Intended Audience :: Developers",
	"Intended Audience :: Education",
	"License :: OSI Approved :: MIT License",
	"Natural Language :: English",
	"Operating System :: OS Independent",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: 3.9",
	"Programming Language :: Python :: Implementation :: CPython",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: Utilities",
	"Topic :: Multimedia :: Graphics"
]
homepage = "https://github.com/FHPythonUtils/PyLottie"
repository = "https://github.com/FHPythonUtils/PyLottie"
documentation = "https://github.com/FHPythonUtils/PyLottie/blob/master/README.md"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "<10,>=8.1.0"
pyppeteer = "<2,>=0.2.4"

[tool.poetry.scripts]
pylottie = 'pylottie:cli'


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
